requestAnimationFrame support
Added a function to detect whether a browser has retina support (for
iOS and new Mac Safari).

Small bug fix to the HtmlParser class.

Made the Stage use the new requestAnimationFrame functionality to
trigger ENTER_FRAME events. This is now also used by the Animation
class so JS animations should now be nicely in sync with CSS
animations, and should be a bit smoother overall. The animation system
(and the ENTER_FRAME events) will automatically fallback to a
setTimeout based loop if the functionality is not available in the
browser.